securewithloadertoaster

+++ src/app/view/secure-pages/login/login.component.ts
@@ -7,49 +7,64 @@ import { ReactiveFormsModule } from '@angular/forms';
 import { AuthService } from '../../../services/auth.service';
 import { RecaptchaModule, RecaptchaFormsModule } from 'ng-recaptcha';
 import { CategoryService } from "../../../services/category.service";
+import { ToastService } from "../../../services/toast.service";
+import { LoaderComponent } from "../../../view/loader/loader.component";
+
 @Component({
   selector: 'app-login',
   standalone: true,
   imports: [RouterLink, CommonModule, ReactiveFormsModule, HttpClientModule, RecaptchaModule,
-    RecaptchaFormsModule,FormsModule],
+    RecaptchaFormsModule,FormsModule,LoaderComponent],
   providers: [AuthService],
   templateUrl: './login.component.html',
   styleUrl: './login.component.css'
 })
 export class LoginComponent {
   fb = inject(FormBuilder);
-  router = inject(Router)
+  router = inject(Router);
   authService = inject(AuthService);
   categoryService = inject(CategoryService);
   loginForm!: FormGroup;
   passwordFieldType: string = 'password';
+  toastService = inject(ToastService);
+  isLoading = false;
+  submitted = false;
 
+  
   ngOnInit(): void {
     this.loginForm = this.fb.group({
       email: ['', Validators.compose([Validators.required, Validators.email])],
       password: ['', Validators.required],
-      recaptcha: [null, Validators.required],
-      rememberMe: [false]
-    })
+      recaptcha: [null, Validators.required],  // Ensuring recaptcha is mandatory
+      rememberMe: [false, Validators.requiredTrue]  // Making rememberMe checkbox mandatory
+    });
   }
 
   togglePasswordVisibility(): void {
     this.passwordFieldType = this.passwordFieldType === 'password' ? 'text' : 'password';
   }
-  resolved(captchaResponse:any) {
+
+  resolved(captchaResponse: any) {
     console.log(`Resolved captcha with response: ${captchaResponse}`);
   }
+
   submit() {
-    console.log(this.loginForm.value);
+    this.submitted = true;
+    this.isLoading = true;  
+    
+    if (this.loginForm.invalid) {
+      this.toastService.show('Failed', 'Please fill in all required fields including the CAPTCHA and Remember Me checkbox');
+      return;
+    }
     this.authService.loginService(this.loginForm.value)
       .subscribe({
         next: (res) => {
-          alert("Login Successfully");
+          this.isLoading = false; 
+          this.toastService.show('Success', 'Login successfully!');
           console.log(res.data.roles);
-  
-          // Always clear sessionStorage to ensure no data remains
-          sessionStorage.clear();
-  
+
+          sessionStorage.clear();  // Always clear sessionStorage
+
           if (this.loginForm.value.rememberMe) {
             localStorage.setItem("user_id", res.data._id);
             localStorage.setItem("token", res.token);
@@ -57,10 +72,9 @@ export class LoginComponent {
             localStorage.setItem("isBuyer", res.data.isBuyer);
             localStorage.setItem("isSeller", res.data.isSeller);
           } else {
-            // Clear localStorage if rememberMe is not checked
             localStorage.clear();
           }
-  
+
           if (res.data.isBuyer) {
             this.router.navigate(['/home']);
             this.categoryService.setHideCategory(false);
@@ -73,8 +87,9 @@ export class LoginComponent {
         },
         error: (err) => {
           console.log(err);
+          this.isLoading = false; 
         }
       });
   }
-  
 }
+
